Added Repositories

+++ Backend/src/main/java/net/ictcampus/waerMachts/model/repositories/InseratRepository.java
@@ -1,8 +1,9 @@
 package net.ictcampus.waerMachts.model.repositories;
 
+import net.ictcampus.waerMachts.model.models.Inserat;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InseratRepository extends CrudRepository<Inserat, Long> {
+public interface InseratRepository extends CrudRepository<Inserat, Integer> {
 }
