+++ Backend/src/main/java/net/ictcampus/waerMachts/model/repositories/SkillRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SkillRepository extends CrudRepository<Skill,Long> {
+public interface SkillRepository extends CrudRepository<Skill,Integer> {
 }
