+++ Backend/src/main/java/net/ictcampus/waerMachts/model/repositories/UserRepository.java
@@ -1,9 +1,9 @@
 package net.ictcampus.waerMachts.model.repositories;
 
-import net.ictcampus.waerMachts.model.User;
+import net.ictcampus.waerMachts.model.models.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<User,Long> {
+public interface UserRepository extends CrudRepository<User,Integer> {
 }
